LayoutTests/svg/dynamic-updates/SVGFEBlendElement-svgdom-in-prop-expected.txt

CONSOLE MESSAGE: line 215: TypeError: 'undefined' is not an object (evaluating 'normalBlend.in1.baseVal = "SourceGraphic"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEBlendElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL normalBlend.in1.baseVal should be SourceAlpha. Threw exception TypeError: 'undefined' is not an object (evaluating 'normalBlend.in1.baseVal')
FAIL multiplyBlend.in1.baseVal should be SourceAlpha. Threw exception TypeError: 'undefined' is not an object (evaluating 'multiplyBlend.in1.baseVal')
FAIL screenBlend.in1.baseVal should be SourceAlpha. Threw exception TypeError: 'undefined' is not an object (evaluating 'screenBlend.in1.baseVal')
FAIL darkenBlend.in1.baseVal should be SourceAlpha. Threw exception TypeError: 'undefined' is not an object (evaluating 'darkenBlend.in1.baseVal')
FAIL lightenBlend.in1.baseVal should be SourceAlpha. Threw exception TypeError: 'undefined' is not an object (evaluating 'lightenBlend.in1.baseVal')
